requirements.txt
Django==1.8.18
django-notifications-hq==1.3
